+++ CHANGES.md
@@ -6,8 +6,9 @@
 - feat: support schema `type: 'file'` with file type
 - feat: `min` and `max` schema validation for Maps
 - feat: add bigint format validation
-- fix: delegate to Zod 4's regexes for string validation
 - feat: validate email string types using any pattern/flags properties
+- feat: add string kinds with validation: jwt (along with optional `algorithm` property), e164, xid, guid, ksuid
+- fix: delegate to Zod 4's regexes for string validation
 - fix: remove dropped `unknownKeys`
 - fix: allow symbol again, but given dropped schema support
 - fix: drop nativeEnum as dropped from Zod/Zodexy